omap: hwmod: add capability to disable usecount tracking on clkdm
For some clockdomains, it does not matter whether some of the hwmods
under them are unidle, they can enter sleep anyway in this case. An
example of such hwmod is GPIO, it does not prevent PER domain from
idling on hardware level. Added software support for tagging these
kinds of hwmods, which ignores the clkdm usecount tracking. Each
individual hwmod which is wanted to behave like this, should be
registered with omap_hwmod_disable_clkdm_usecounting().

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>