ktl/check_component.py: Remove linux-meta-ti-omap4 main enforce
linux-tools-omap4 from linux-meta-ti-omap4 ended in universe for release
pocket on some distros, and components after release are not changed
usually to avoid issues (like the package might have a dependency on
another universe package). Remove enforce of linux-meta-ti-omap4 to be
on main, to fall back to the default check (that uses what's on release
pocket by default as a source of information), so the checker doesn't
complain about linux-tools-omap4 being on universe.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>