stable/sru-workflow-manager: wait 1 hour after promote-to-proposed is done
Usually the archive admins set the promote-to-proposed task to Fix
Released right away after doing the copy, but actual publishing takes
more time. It's said that the publishing should take at most 1 hour to
complete, but could be a bit more in some cases, or could vary. In any
case, wait arbitrarily 1 hour before we do the actual processing of this
state, to avoid sru-workflow-manager to complain about packages in wrong
component while real copying/publishing still didn't finish, in the case
the packages were previously in the wrong component and are now moved to
the right ones.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>