stable/apply-stable-patches: new script to help with stable maintenance
This new script is a experiment intended to help with 3.2 stable kernel
maintenance. For example, on a linux 3.2 stable branch, this would be
the use of this script:

cd linux-stable
git checkout -b linux-3.2.y -t origin/linux-3.2.y
git clean -dxf
./stable/apply-stable-patches \
	--range=c9651e70ad0aa499814817cbf3cc1d0b806ed3a1..FETCH_HEAD \
	--sob="Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>"

When running, apply-stable-patches will search for all patches marked
for stable in the upstream range given (in the example above,
c9651e70ad0aa499814817cbf3cc1d0b806ed3a1..FETCH_HEAD, FETCH_HEAD is the
upstream HEAD, the script fetches directly from master Linus tree).

Then, it will try to apply all patches marked for stable for 3.2.
Succesfuly applied patches will be added to the current branch
(linux-3.2.y). Status of application of the patches marked for stable
are displayed on the screen.

Also after running, 5 directories should be created, and patches are
copied to them accordingly to the application status:
Applied - successfuly applied patches
Failed - patches which failed to apply
Ignored - patches not applied because it's Cc: line told it was for a
          different version than we are trying to apply
Discarded - nothing copied by the script to it, user will copy here
            patches it want to exclude on next run
Fixed - nothing copied here by the script, user will move fixed patches
        from Failed here

Then while we have patches in Failed (1), we must:
- fix the patch, and move it to Fixed directory, keeping the same file
  name
- drop the patch, moving to Discarded, keeping the same file name
And run:
* git reset --hard origin/linux-3.2.y
* ./stable/apply-stable-patches \
        --range=c9651e70ad0aa499814817cbf3cc1d0b806ed3a1..FETCH_HEAD \
        --sob="Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>"
If there are new failed patches because of patch ordering issues etc.,
go back to (1)

After there is no more patches in Failed directory, we should review
all patches in Ignored directory. In case the patch shouldn't be
ignored, and must be applied to 3.2, move to Fixed (making sure the
patch applies correctly). If we moved any patch to Fixed, we should do
again:
* git reset --hard origin/linux-3.2.y
* ./stable/apply-stable-patches \
        --range=c9651e70ad0aa499814817cbf3cc1d0b806ed3a1..FETCH_HEAD \
        --sob="Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>"

Then as a last step, reset and reapply the changes:
* git reset --hard origin/linux-3.2.y
* ./stable/apply-stable-patches \
        --range=c9651e70ad0aa499814817cbf3cc1d0b806ed3a1..FETCH_HEAD \
        --sob="Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>"

If nothing fails (we applied all necessary patches, and discarded and
ignored what is needed), then the branch is ready to be pushed.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>